saveing before checking out

+++ setup.sh
@@ -2,61 +2,60 @@
 #################################################
 # This script should be run after fresh install #
 #################################################
-# variables
-PREFIX_=$( cd "$( dirname "$0" )" >/dev/null 2>&1 && pwd )
-toInstall=()
-
-# functions
-source $PREFIX_/functions.sh
+# variables & functions
+source $( cd "$( dirname "$0" )" >/dev/null 2>&1 && pwd )/functions.sh
 
 ##########################
 # Start installing stuff #
 ##########################
 title "Setting up Linux for $OS"
 title "Starting base setup"
-FASE="base"
-if [[ $OS != "NAS" ]]; then
-    # For Debian(-like) systems
-    subtitle "Change password for $USER"
-    sudo passwd $USER
+if [[ -f $PREFIX_/.base_done ]]; then
+    echo "Base done"
+else
+    FASE="base"
+    if [[ $OS != "NAS" ]]; then
+        # For Debian(-like) systems
+        subtitle "Change password for $USER"
+        sudo passwd $USER
 
-    if [[ $OS == "ChromeOS" ]]; then
-        subtitle "Upgrade container"
-        sudo bash /opt/google/cros-containers/bin/upgrade_container
-    else
-        subsubtitle "Updating system"
-        sudo apt-get update -q
-        sudo apt-get upgrade -qq
-        sudo apt-get dist-upgrade -y
+        if [[ $OS == "ChromeOS" ]]; then
+            subtitle "Upgrade container"
+            sudo bash /opt/google/cros-containers/bin/upgrade_container
+        else
+            subsubtitle "Updating system"
+            sudo apt-get update -q
+            sudo apt-get upgrade -qq
+            sudo apt-get dist-upgrade -y
+        fi
+        #sudo apt autoremove --purge -qq
+        
+        subtitle "Writing list of base-packages to ~/.base-packages"
+        sudo apt list --installed > ~/.base-packages
+        
+        subtitle "Adding rules to apt.conf.d for minimal installation packages"
+        cpfile 99_norecommends /etc/apt/apt.conf.d
+        
+        subtitle "Installing base-packages"
+        #sudo apt --fix-broken install -y
+        sudo apt-get install -qq nano #git
+        
+        if [[ $OS == "ChromeOS" ]]; then
+            sudo apt-get -qq install mesa-utils
+        elif [[ $OS == "OSMC" ]]; then
+            echo "No specific additional packages specified for $OS packages"
+        fi
+    elif $OS == "NAS"; then
+        # For Synology DSM
+        echo "no initial setup defined for $OS: nothing to do"
     fi
-    #sudo apt autoremove --purge -qq
-    
-    subtitle "Writing list of base-packages to ~/.base-packages"
-    sudo apt list --installed > ~/.base-packages
-    
-    subtitle "Adding rules to apt.conf.d for minimal installation packages"
-    cpfile 99_norecommends /etc/apt/apt.conf.d
-    
-    subtitle "Installing base-packages"
-    #sudo apt --fix-broken install -y
-    sudo apt-get install -qq nano #git
-    
-    if [[ $OS == "ChromeOS" ]]; then
-        sudo apt-get -qq install mesa-utils
-    elif [[ $OS == "OSMC" ]]; then
-        echo "No specific additional packages specified for $OS packages"
-    fi
-elif $OS == "NAS"; then
-    # For Synology DSM
-    echo "no initial setup defined for $OS: nothing to do"
+
+    # setting up git
+    [[ $(git config --get user.email) == "" ]] && read -p "Enter e-mail adres for git [you@example.com]: " gite && git config --global user.email "$gite"
+    [[ $(git config --get user.name) == "" ]] && read -p "Enter name for git [John Doe]: " gitn && git config --global user.name "$gitn"
+
+    echo "--- Base setup DONE ---"
 fi
-
-# setting up git
-[[ $(git config --get user.email) == "" ]] && read -p "Enter e-mail adres for git [you@example.com]: " gite && git config --global user.email "$gite"
-[[ $(git config --get user.name) == "" ]] && read -p "Enter name for git [John Doe]: " gitn && git config --global user.name "$gitn"
-
-echo "--- Base setup DONE ---"
-
 ###########
 # PACKGES #
 ###########
